+++ integration_tests/tests/flowmachine_server_tests/test_server.py
@@ -53,6 +53,7 @@ def test_get_available_queries(send_zmq_message_and_receive_reply):
                 "unique_subscriber_counts",
                 "location_introversion",
                 "total_network_objects",
+                "aggregate_network_objects",
                 "dfs_metric_total_amount",
                 "spatial_aggregate",
                 "joined_spatial_aggregate",